Split 'part' model into new file
